Ability to add/modify scales in dashboard
Used the existing minWords/maxWords fields for the purpose of min/max value on the scale (maybe more logical to add fields later on). Similarly failure / success message are used
for labels for the maxima of the scale, since they have no other use in scale items.

Changes in scale item types are not yet reflected in the widgets, which still assume the standard 7-step scale.