Fixed invalid hook error and updated QuizImpl to function component
Even the right usage of hooks in components lead to invalid hook error https://reactjs.org/warnings/invalid-hook-call-warning.html.
The cause was that example and widgets used different react modules. Explicitly using the react module
in the node_modules of the parent directory with the example fixed the issue, but is not the prettiest way...