Added the status of feedback displaying to the redux store
When loading the quiz, the feedback is displayed only if the user quiz state is locked. After submitting, the feedback is displayed only
if the user quiz state becomes locked or the user has already received all the points in the quiz. If the feedback is shown
and the user changes the value of a feedbacked item, all the feedback messages are immediately hidden.

At least everything not involving peer reviews should now support this system.