Clearer state after submitting incorrect answer, tries remaining
Added option to display message until the first change after a submit attemp. This is used when displaying the
message that the answer has been submitted, but it was not correct. Also, the button only displays 'submitted',
not 'submitted successfully' (and the same in Finnish).

At some point this functionality will have to be placed inside the reducers of the store, not action creators...